# microservices/kafka_consumer/app/udaconnect/schemas.py
from app.udaconnect.models import Location
from marshmallow import Schema, fields


class LocationSchema(Schema):
    id = fields.Integer()
    person_id = fields.Integer()
    longitude = fields.String(attribute="longitude")
    latitude = fields.String(attribute="latitude")
    creation_time = fields.DateTime()

    class Meta:
        model = Location


# class PersonSchema(Schema):
#     id = fields.Integer()
#     first_name = fields.String()
#     last_name = fields.String()
#     company_name = fields.String()

#     class Meta:
#         model = Person


# class ConnectionSchema(Schema):
#     location = fields.Nested(LocationSchema)
#     person = fields.Nested(PersonSchema)